Scope registry working
