+++ src/MeshWeaver.BusinessRules/IScopeRegistryFactory.cs
@@ -1,24 +1,7 @@
 ﻿using System.Collections.Immutable;
 using System.Reflection;
-using MeshWeaver.Messaging;
-using Microsoft.Extensions.DependencyInjection;
 
 namespace MeshWeaver.BusinessRules;
-
-public static class BusinessRulesRegistryExtensions
-{
-    public static MessageHubConfiguration AddBusinessRules(this MessageHubConfiguration configuration, params Assembly[] assemblies)
-    => configuration.WithServices(services =>
-            services
-                .AddScoped<IScopeRegistryFactory, ScopeRegistryFactory>()
-                .AddScoped<IScopeFactory>(sp => new ScopeFactory(assemblies))
-    );
-
-    public static ScopeRegistry<TState> GetScopeRegistry<TState>(this IMessageHub hub, TState state)
-    {
-        return hub.ServiceProvider.GetRequiredService<IScopeRegistryFactory>().Create(state);
-    }
-}
 
 public interface IScopeRegistryFactory
 {
@@ -27,17 +10,17 @@ public interface IScopeRegistryFactory
 
 public interface IScopeFactory
 {
-    object Create<TScope, TState>(object identity, TState state);
+    object Create<TScope, TState>(object identity, ScopeRegistry<TState> state);
 }
 
-public class ScopeFactory(Assembly[] assemblies) : IScopeFactory
+internal class ScopeFactory(Assembly[] assemblies) : IScopeFactory
 {
     private readonly IReadOnlyDictionary<Type, Type> scopeImplementations = assemblies
         .SelectMany(a => a.GetTypes()
             .Where(t => !t.IsAbstract && t.BaseType?.IsGenericType == true && t.BaseType.GetGenericTypeDefinition() == typeof(ScopeBase<,,>))
         )
         .ToDictionary(t => t.BaseType!.GetGenericArguments().First());
-    public object Create<TScope, TState>(object identity, TState state)
+    public object Create<TScope, TState>(object identity, ScopeRegistry<TState> state)
     {
         if (!scopeImplementations.TryGetValue(typeof(TScope), out var type))
             throw new NotSupportedException($"No implementation found for type {typeof(TScope).Name}. Is the compiler installed?");
@@ -45,7 +28,7 @@ public class ScopeFactory(Assembly[] assemblies) : IScopeFactory
     }
 }
 
-public class ScopeRegistryFactory(IScopeFactory scopeFactory) : IScopeRegistryFactory
+internal class ScopeRegistryFactory(IScopeFactory scopeFactory) : IScopeRegistryFactory
 {
     public ScopeRegistry<TState> Create<TState>(TState state)
     {
@@ -53,7 +36,7 @@ public class ScopeRegistryFactory(IScopeFactory scopeFactory) : IScopeRegistryFa
     }
 }
 
-public partial class ScopeRegistry<TState>(TState state, IScopeFactory scopeFactory)
+public class ScopeRegistry<TState>(TState state, IScopeFactory scopeFactory)
 {
     private ImmutableDictionary<(Type ScopeType, object Identity), object> scopes = ImmutableDictionary<(Type ScopeType, object Identity), object>.Empty;
     private readonly object locker = new();
@@ -67,7 +50,7 @@ public partial class ScopeRegistry<TState>(TState state, IScopeFactory scopeFact
             {
                 return (TScope)scope;
             }
-            var newScope = scopeFactory.Create<TScope, TState>(identity, state);
+            var newScope = scopeFactory.Create<TScope, TState>(identity, this);
             scopes = scopes.Add(key, newScope);
             return (TScope)newScope;
